fixed task 0

+++ 0x13-more_singly_linked_lists/0-print_listint.c
@@ -6,7 +6,7 @@
 */
 size_t print_listint(const listint_t *h)
 {
-	int count_node = 0;
+	size_t count_node = 0;
 
 	if (h == NULL)
 		return (0);
